Firmware/4.x/winter_scripts/requirements.txt
picamera2
numpy
opencv-python
Pillow
RPi.GPIO
piexif
python-crontab
schedule
